refactor: Reservation 엔티티와 연관 타입을 별도의 도메인 패키지로 이동
