+++ src/main/java/com/jvnlee/catchdining/domain/reservation/model/ReservationStatus.java
@@ -1,4 +1,4 @@
-package com.jvnlee.catchdining.entity;
+package com.jvnlee.catchdining.domain.reservation.model;
 
 public enum ReservationStatus {
     RESERVED, VISITED, NO_SHOW, CANCEL
