+++ src/main/java/com/jvnlee/catchdining/domain/user/model/User.java
@@ -1,5 +1,6 @@
 package com.jvnlee.catchdining.domain.user.model;
 
+import com.jvnlee.catchdining.domain.reservation.model.Reservation;
 import com.jvnlee.catchdining.domain.user.dto.UserDto;
 import com.jvnlee.catchdining.entity.*;
 import lombok.Getter;
